# src/main/resources/static/doc/openapi.json
{
  "components": {
    "schemas": {
      "ResponseResultOrderAndProductInfoDto_response": {
        "properties": {
          "code": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer",
            "example": "900"
          },
          "message": {
            "format": "string",
            "description": "No comments found.",
            "type": "string",
            "example": "success"
          },
          "data": {
            "description": "No comments found.(object)",
            "type": "object",
            "$ref": "#/components/schemas/OrderAndProductInfoDto_response"
          }
        }
      },
      "string": {
        "format": "string",
        "type": "string"
      },
      "OrderAndProductInfoDto_response": {
        "properties": {
          "orderId": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": "765"
          },
          "orderTitle": {
            "format": "string",
            "description": "No comments found.",
            "type": "string",
            "example": "y3bia8"
          },
          "goodsId": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": "776"
          },
          "productNum": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer",
            "example": "213"
          },
          "productLastNum": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer",
            "example": "44"
          }
        }
      },
      "ResponseResultString_response": {
        "properties": {
          "code": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer",
            "example": "885"
          },
          "message": {
            "format": "string",
            "description": "No comments found.",
            "type": "string",
            "example": "success"
          },
          "data": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": ""
    }
  ],
  "openapi": "3.0.3",
  "paths": {
    "/order/v1/domainEvent/{orderId}": {
      "get": {
        "summary": "接入层响应外部请求",
        "deprecated": false,
        "description": "接入层响应外部请求",
        "tags": [
          "接入层"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResultString_response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"code\": 877,\n  \"message\": \"success\",\n  \"data\": \"ar1gj0\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "orderv1domainEvent{orderId}",
        "parameters": [
          {
            "schema": {
              "format": "integer",
              "type": "int64"
            },
            "in": "path",
            "description": "No comments found.",
            "required": true,
            "name": "orderId"
          }
        ]
      }
    },
    "/order/v1/getOrderInfo/{orderId}": {
      "get": {
        "summary": "接入层响应外部请求\r\n同时查询返回两个Domain下的数据",
        "deprecated": false,
        "description": "接入层响应外部请求\r\n同时查询返回两个Domain下的数据",
        "tags": [
          "接入层"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResultOrderAndProductInfoDto_response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"code\": 668,\n  \"message\": \"success\",\n  \"data\": {\n    \"orderId\": 587,\n    \"orderTitle\": \"qwlrz6\",\n    \"goodsId\": 714,\n    \"productNum\": 193,\n    \"productLastNum\": 869\n  }\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "orderv1getOrderInfo{orderId}",
        "parameters": [
          {
            "schema": {
              "format": "integer",
              "type": "string"
            },
            "in": "path",
            "description": "String",
            "required": true,
            "name": "orderId"
          }
        ]
      }
    },
    "/order/v1/domainLogicalProcessing/{orderId}": {
      "get": {
        "summary": "接入层响应外部请求",
        "deprecated": false,
        "description": "接入层响应外部请求",
        "tags": [
          "接入层"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResultString_response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"code\": 32,\n  \"message\": \"success\",\n  \"data\": \"yuc01l\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "orderv1domainLogicalProcessing{orderId}",
        "parameters": [
          {
            "schema": {
              "format": "integer",
              "type": "int64"
            },
            "in": "path",
            "description": "No comments found.",
            "required": true,
            "name": "orderId"
          }
        ]
      }
    },
    "/order/v1/testCGLIB/{orderId}": {
      "get": {
        "summary": "接入层响应外部请求",
        "deprecated": false,
        "description": "接入层响应外部请求",
        "tags": [
          "接入层"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResultString_response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"code\": 566,\n  \"message\": \"success\",\n  \"data\": \"ilqmpq\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "orderv1testCGLIB{orderId}",
        "parameters": [
          {
            "schema": {
              "format": "integer",
              "type": "int64"
            },
            "in": "path",
            "description": "No comments found.",
            "required": true,
            "name": "orderId"
          }
        ]
      }
    }
  },
  "info": {
    "title": "cnaworld-base",
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "接入层",
      "description": "接入层"
    }
  ]
}